requirements.txt
cycler==0.11.0
fonttools==4.37.1
kiwisolver==1.4.4
matplotlib==3.5.3
neat-python==0.92
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pip==22.2.2
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
setuptools==60.2.0
six==1.16.0
visualize==0.5.1
wheel==0.37.1
